✨ feat(format-mjs): Add code to retrieve values from plugin
This commit adds a new block of code to format.mjs that loops through
a list of selectors and retrieves values from a plugin. If the retrieved
values are not empty, they are added to a "rawData" object. The condition
for executing this block of code is if the "rawData" object is not empty.

The changes are done to enhance the functionality of format.mjs by
allowing it to retrieve values from a plugin and add them to the "rawData"
object if they are not empty.